+++ e2e-tests/src/lib.rs
@@ -4,4 +4,3 @@ pub mod config;
 pub mod staking;
 pub mod test;
 pub mod transfer;
-mod waiting;
