+++ e2e-tests/src/test/finalization.rs
@@ -1,7 +1,6 @@
-use aleph_client::create_connection;
+use aleph_client::{create_connection, wait_for_finalized_block};
 
 use crate::config::Config;
-use crate::waiting::wait_for_finalized_block;
 
 pub fn finalization(config: &Config) -> anyhow::Result<u32> {
     let connection = create_connection(&config.node, config.protocol);
