--- e2e-tests/src/waiting.rs
@@ -1,22 +1,0 @@
-use std::sync::mpsc::channel;
-use log::info;
-use aleph_client::{Connection, Header};
-
-/// blocks the main thread waiting for a block with a number at least `block_number`
-pub fn wait_for_finalized_block(connection: &Connection, block_number: u32) -> anyhow::Result<u32> {
-    let (sender, receiver) = channel();
-    connection.subscribe_finalized_heads(sender)?;
-
-    while let Ok(header) = receiver
-        .recv()
-        .map(|h| serde_json::from_str::<Header>(&h).unwrap())
-    {
-        info!("[+] Received header for a block number {:?}", header.number);
-
-        if header.number.ge(&block_number) {
-            return Ok(block_number);
-        }
-    }
-
-    Err(anyhow::anyhow!("Giving up"))
-}
